+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentIT.scala
@@ -7,7 +7,6 @@
 package edu.ie3.simona.agent.em
 
 import edu.ie3.datamodel.models.result.system.EmResult
-import edu.ie3.simona.agent.grid.GridAgent
 import edu.ie3.simona.agent.participant.ParticipantAgentInit
 import edu.ie3.simona.agent.participant.ParticipantAgentInit.{
   ParticipantRefs,
@@ -91,7 +90,6 @@ class EmAgentIT
   "An em agent" when {
     "having load, pv and storage agents connected" should {
       "be initialized correctly and run through some activations" in {
-        val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultServiceProxy =
           TestProbe[ResultEvent | ExpectResult]("ResultServiceProxy")
         val primaryServiceProxy =
@@ -100,7 +98,6 @@ class EmAgentIT
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
-          gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryServiceProxy.ref,
           resultServiceProxy = resultServiceProxy.ref,
           services = Map(ServiceType.WeatherService -> weatherService.ref),
@@ -378,7 +375,6 @@ class EmAgentIT
 
     "having load, pv and heat pump agents connected" should {
       "be initialized correctly and run through some activations" in {
-        val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultServiceProxy =
           TestProbe[ResultEvent | ExpectResult]("ResultServiceProxy")
         val primaryServiceProxy =
@@ -387,7 +383,6 @@ class EmAgentIT
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
-          gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryServiceProxy.ref,
           resultServiceProxy = resultServiceProxy.ref,
           services = Map(ServiceType.WeatherService -> weatherService.ref),
@@ -946,7 +941,6 @@ class EmAgentIT
 
     "having a pv and a load agent connected" should {
       "have correct values also for agents with limited operation time" in {
-        val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultServiceProxy =
           TestProbe[ResultEvent | ExpectResult]("ResultServiceProxy")
         val primaryServiceProxy =
@@ -955,7 +949,6 @@ class EmAgentIT
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
-          gridAgent = gridAgent.ref,
           primaryServiceProxy = primaryServiceProxy.ref,
           resultServiceProxy = resultServiceProxy.ref,
           services = Map(ServiceType.WeatherService -> weatherService.ref),