+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -116,7 +116,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map.empty
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -280,7 +279,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map.empty
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -441,7 +439,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map(service.ref -> 0)
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -725,7 +722,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map(service.ref -> 0)
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -961,7 +957,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map.empty
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -1166,7 +1161,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map.empty
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -1433,7 +1427,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map(service.ref -> 0)
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),
@@ -1876,7 +1869,6 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
               Map(service.ref -> 0)
             ),
             ParticipantGridAdapter(
-              gridAgent.ref,
               expectedRequestTick = 12 * 3600,
               requestVoltageDeviationTolerance = Each(1e-14),
             ),